requirements-dev.txt
# Development dependencies
pytest>=7.0
pytest-django>=4.5
black>=23.0
isort>=5.12
flake8>=6.0
coverage>=7.0

# Documentation
sphinx>=5.0
sphinx-rtd-theme>=1.2
